+++ Astral One Engine/Engine/Core/Map Features/Terrain.swift
@@ -2,12 +2,12 @@ import Foundation
 
 public class Terrain {
     public let name: String
-    public let food: Float
-    public let shields: Float
-    public let trade: Float
-    public let movementCost: Float
+    public let food: Double
+    public let shields: Double
+    public let trade: Double
+    public let movementCost: Double
     
-    public init(name: String, food: Float, shields: Float, trade: Float, movementCost: Float) {
+    public init(name: String, food: Double, shields: Double, trade: Double, movementCost: Double) {
         self.name = name
         self.food = food
         self.shields = shields
